refactor(client): refactor client file structor
refactor some client file structors for better understanding
