+++ packages/client/src/api/index.ts
@@ -1,0 +1,2 @@
+export * from './discover';
+export * from './grant-token';
