+++ packages/client/src/modules/errors.ts
