+++ packages/client/src/modules/session-manager.test.ts
@@ -1,4 +1,4 @@
-import { SESSION_MANAGER_KEY } from './constants';
+import { SESSION_MANAGER_KEY } from '../constants';
 import SessionManager, { Session } from './session-manager';
 import { SessionStorage } from './storage';
