+++ packages/client/src/utils/assembler.ts
@@ -1,6 +1,6 @@
 import qs from 'query-string';
 
-import { generateCodeChallenge, generateCodeVerifier, generateState } from './generators';
+import { generateCodeVerifier, generateCodeChallenge, generateState } from './generators';
 
 export interface LoginPrepareParameters {
   baseUrl: string;
@@ -9,6 +9,16 @@ export interface LoginPrepareParameters {
   redirectUri: string;
 }
 
+/**
+ * Generate loginUrl
+ *
+ * @param {Object} parameters
+ * @param {String} parameters.baseUrl
+ * @param {String} parameters.clientId
+ * @param {String} [parameters.scope]
+ * @param {String} [parameters.redirectUri]
+ * @returns
+ */
 export const getLoginUrlWithCodeVerifierAndState = async (
   parameters: LoginPrepareParameters
 ): Promise<{ url: string; codeVerifier: string; state: string }> => {
@@ -29,4 +39,20 @@ export const getLoginUrlWithCodeVerifierAndState = async (
   })}`;
 
   return { url, codeVerifier, state };
+};
+
+/**
+ * Generate Logout Url
+ * @param {String} baseUrl
+ * @param {String} idToken
+ * @param {String} redirectUri
+ * @returns
+ */
+export const getLogoutUrl = (baseUrl: string, idToken: string, redirectUri: string) => {
+  const url = `${baseUrl}?${qs.stringify({
+    id_token_hint: idToken,
+    post_logout_redirect_uri: redirectUri,
+  })}`;
+
+  return url;
 };
