+++ packages/client/src/utils/id-token.test.ts
@@ -2,8 +2,9 @@ import { KeyObject } from 'crypto';
 
 import { SignJWT, generateKeyPair } from 'jose';
 import nock from 'nock';
+import { StructError } from 'superstruct';
 
-import { createJWKS, verifyIdToken } from './verify-id-token';
+import { decodeToken, createJWKS, verifyIdToken } from './id-token';
 
 describe('verifyIdToken', () => {
   test('valid idToken', async () => {
@@ -103,5 +104,35 @@ describe('verifyIdToken', () => {
     await expect(verifyIdToken(JWKS, idToken, 'foo')).rejects.toThrowError(
       '"exp" claim timestamp check failed'
     );
+  });
+});
+
+describe('decodeToken', () => {
+  test('decode token and get claims', async () => {
+    const JWT = await new SignJWT({})
+      .setProtectedHeader({ alg: 'RS256' })
+      .setAudience('foo')
+      .setSubject('foz')
+      .setIssuer('logto')
+      .setIssuedAt()
+      .setExpirationTime('2h')
+      .sign((await generateKeyPair('RS256')).privateKey);
+    const payload = decodeToken(JWT);
+    expect(payload.sub).toEqual('foz');
+  });
+
+  test('throw on invalid JWT string', async () => {
+    expect(() => decodeToken('invalid-JWT')).toThrow();
+  });
+
+  test('throw ZodError when iss is missing', async () => {
+    const JWT = await new SignJWT({})
+      .setProtectedHeader({ alg: 'RS256' })
+      .setAudience('foo')
+      .setSubject('foz')
+      .setIssuedAt()
+      .setExpirationTime('2h')
+      .sign((await generateKeyPair('RS256')).privateKey);
+    expect(() => decodeToken(JWT)).toThrowError(StructError);
   });
 });
